+++ src/app/app.module.ts
@@ -7,6 +7,7 @@ import {HeaderComponent} from './layout/header/header.component';
 import {AuthInterceptor} from "./http/auth.interceptor";
 import {EntrancePageModule} from "./movie/pages/entrance-page/entrance-page.module";
 import {MovieInnerPageModule} from "./movie/pages/movie-inner-page/movie-inner-page.module";
+import {UrlInterceptor} from "./http/url.interceptor";
 
 @NgModule({
   declarations: [
@@ -20,11 +21,17 @@ import {MovieInnerPageModule} from "./movie/pages/movie-inner-page/movie-inner-p
     AppRoutingModule,
     HttpClientModule,
   ],
-  providers: [{
-    provide: HTTP_INTERCEPTORS,
-    useClass: AuthInterceptor,
-    multi: true,
-  },],
+  providers: [
+    {
+      provide: HTTP_INTERCEPTORS,
+      useClass: AuthInterceptor,
+      multi: true,
+    },
+    {
+      provide: HTTP_INTERCEPTORS,
+      useClass: UrlInterceptor,
+      multi: true,
+    },],
   bootstrap: [AppComponent]
 })
 export class AppModule {
