+++ src/app/http/url.interceptor.ts
@@ -1,0 +1,14 @@
+import {HttpEvent, HttpHandler, HttpInterceptor, HttpRequest} from "@angular/common/http";
+import {Injectable} from "@angular/core";
+import {Observable} from "rxjs";
+
+const BASE_URL = 'https://api.themoviedb.org/3';
+
+@Injectable()
+export class UrlInterceptor implements HttpInterceptor {
+  intercept(req: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
+    return next.handle(req.clone({
+      url: BASE_URL + req.url
+    }));
+  }
+}
